Fix sans-serif font stack on 'CSS Font Styles' slide
The font stack is all sans-serif fonts. Having
the final font in the stack be a serif font
makes it difficult for students to understand
and could potentially lead the student into
mixing up "sans-serif" and "serif" fonts. This
fix reenforces the differences between
"sans-serif" and "serif" by demonstrating that
Arial and Helvetic don't have serifs.